Capitalize first letter

+++ app/models/receipt.rb
@@ -6,6 +6,6 @@ class Receipt < ActiveRecord::Base
   accepts_nested_attributes_for :receipts_fees_heads, allow_destroy: true
 
   def total
-    receipts_fees_heads.sum(:amount).to_i
+    receipts_fees_heads.sum(:amount).to_f
   end
 end
